+++ dxHelloworld1.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "dxHelloworld1", "dxHelloworld1\dxHelloworld1.vcxproj", "{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}"
 EndProject
@@ -15,16 +15,16 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Debug|x64.ActiveCfg = Debug|Win32
-		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Debug|x64.Build.0 = Debug|Win32
+		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Debug|x64.ActiveCfg = Debug|x64
+		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Debug|x64.Build.0 = Debug|x64
 		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Debug|x86.ActiveCfg = Debug|Win32
 		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Debug|x86.Build.0 = Debug|Win32
 		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Release|x64.ActiveCfg = Release|x64
 		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Release|x64.Build.0 = Release|x64
 		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Release|x86.ActiveCfg = Release|Win32
 		{5BAEF92D-D7CF-43C6-BB0A-D10A963F4DF8}.Release|x86.Build.0 = Release|Win32
-		{E0B52AE7-E160-4D32-BF3F-910B785E5A8E}.Debug|x64.ActiveCfg = Debug|Win32
-		{E0B52AE7-E160-4D32-BF3F-910B785E5A8E}.Debug|x64.Build.0 = Debug|Win32
+		{E0B52AE7-E160-4D32-BF3F-910B785E5A8E}.Debug|x64.ActiveCfg = Debug|x64
+		{E0B52AE7-E160-4D32-BF3F-910B785E5A8E}.Debug|x64.Build.0 = Debug|x64
 		{E0B52AE7-E160-4D32-BF3F-910B785E5A8E}.Debug|x86.ActiveCfg = Debug|Win32
 		{E0B52AE7-E160-4D32-BF3F-910B785E5A8E}.Debug|x86.Build.0 = Debug|Win32
 		{E0B52AE7-E160-4D32-BF3F-910B785E5A8E}.Release|x64.ActiveCfg = Release|x64
